move package-components to it's own file
